Improved the prompt size, added some metadata
The prompt size was too small. I thought that passing more
than 100 characters to StableDiffusion was pretty much useless,
but turns out that I was wrong.
You can actually pass pretty large strings to StableDiffusers.
So the limit was upped to 500 characters.

Also the PNG metadata didn't contain the diffusers version,
and that's actually a serious problem since recent version of
diffusers outputs seem slightly different, making it almost
impossible to get the exact same result.
So the diffusers version was added, in order to understand
which version is being used.

Signed-off-by: Voyage <voyage@miouyouyou.fr>