Preparing for a Docker release
In order to run the bot inside a Docker, I decided to add a
STABLEDIFFUSION_CACHE_DIR which is passed to StableDiffusionPipeline.
That way, when using Docker, it's possible to mount a volume
and set the cache dirpath to the volume dirpath, and avoid
redownloading all the files on each launch.

I also added a few more variables :

* STABLEDIFFUSION_LOCAL_ONLY (Default to false)
* STABLE_DIFFUSION_MODEL_NAME (Default to CompVis/stable-diffusion-v1-4)
* TORCH_DEVICE (Default to cuda)

I need to update the README.md on these parts but :
* HUGGINGFACES_TOKEN is not required anymore when
  STABLEDIFFUSION_LOCAL_ONLY is set.

* DISCORD_GUILD_ID is now ignored, the commands are reproduced on every
  server the bot is connected to.
  I'll reintroduce some filtering system later, but at the moment, this
  should do the trick.

Signed-off-by: Voyage <voyage@miouyouyou.fr>