Creating config/ directory and moving replacers.json into it
The idea behind this is the ability to configure the bot easily
when using Docker.
Currently, only replacers.json is placed in config/, but others
files might get added in the near future, in order to deal with
words denied in prompts, and witty lines for NSFW results.

Also, switched from print to logger on StableDiffusion side.
I still have to test its effects with Docker, but this might
help avoiding some kludge from Docker's side, in which case I
might release the Docker installation today, and proceed to
Virtual Enivronments tomorrow.

Signed-off-by: Voyage <voyage@miouyouyou.fr>