Added prompt filtering through config/denied_expressions.txt
If config/denied_expression.txt is found, each line of this
file is interpreted as a Regular Expression and used to
analyse prompts generations.

If any of these lines match the prompt of a request, the
request is immediately denied.

This allows bot runners to avoid dealing with unwanted
generations.

If the file is not present, the array of denied expressions
will be empty, and so no prompt filtering will happen, which
is the default behaviour.

Signed-off-by: Voyage <voyage@miouyouyou.fr>