Working on Linux, and with Waifu Diffusion !
The installation is now tested on Linux, through Docker-Compose
and Virtual Environments.

And not only has it been tested on Linux, but I also tested
changing STABLEDIFFUSION_MODEL_NAME to hakurei/waifu-diffusion,
and adding STABLEDIFFUSION_MODE=fp16 on top of that, and it
WORKED OUT OF THE BOX !

YAAAAY !

I also made sure that the venv script work on Linux.
The new bash script also works on Windows.

A few last tests and information on how to setup
STABLEDIFFUSION_MODEL_NAME in the README, and this will be it !

Signed-off-by: Voyage <voyage@miouyouyou.fr>